BF: do not use date and -n1 with rev-list -- get ones directly from the candidates[0]
As I have mentioned before, not sure if 'date' criteria would be the most robust here,
and decided to just use the actual candidate, the child of which we should take.

With -n1, if there are multiple commits, we would be getting the latest on the HEAD
candidate, and not the one at the bottom.  As -n1 is applied before --reverse, --reverse
is of no help.  So I decided that we should just read all, split, and take the last one